fix the selected index disappearing from the select box
The reason this was happening was that whenever the indices variable was
updated the whole select and ng-options values were updated and for some
reason when that happened it ignores the model you have selected. So if
you had a 3s autorefresh it would take 3s to disappear, or 5s or 1min...
The way this fix works is by adding a default value to the ng-options.

Really this is just visual because the ng-model was never lost so you
could still select from the rest of selects.